pyproject.toml
[project]
name = "uav-system"
version = "0.1.0"
description = "Constraint-Aware Fault-Tolerant Multi-Agent UAV System with OODA Loop"
readme = "README.md"
requires-python = ">=3.8"
authors = [
    { name = "Vítor Eulálio Reis", email = "vitor.ereis@proton.me" }
]
keywords = ["uav", "multi-agent", "ooda-loop", "fault-tolerant", "autonomous-systems"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "numpy>=1.24.0",
    "scipy>=1.10.0",
    "matplotlib>=3.7.0",
    "networkx>=3.0",
    "pyyaml>=6.0",
    "jsonschema>=4.17.0",
    "pandas>=2.0.0",
    "flask>=2.3.0",
    "flask-socketio>=5.3.0",
    "python-socketio>=5.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
uav-gcs = "gcs.main:main"
uav-launch = "launch:main"
uav-dashboard = "run_dashboard:main"
uav-launch-gui = "launch_with_gui:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["uav", "gcs", "visualization", "missions"]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
]

[tool.ruff]
line-length = 100
target-version = "py38"

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
ignore = ["E501"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]

max-complexity = 10

# List of directories/files to exclude
exclude = [
    ".git",
    "__pycache__",
    "docs",
    "venv",          # Virtual environments
    ".venv",         # Another common name for virtual environments
    "migrations/*",  # Database migration files (often auto-generated)
    "build/*",       # Build directories
    "setup.py",      # Often excluded as it is treated differently
    ".venv",
    "venv",
    "migrations/*",
    "build/*",
]

# List of error codes to ignore (common ones for use with Black)
ignore = [
    "W503",  # Line break before binary operator
    "E203",  # Whitespace before ':'
    "E501",  # line too long (81 > 79 characters)
]

